Add update-builds action
This action allows us to update a DynamoDB table with build information.
We use this table to help us generate our release tags, so need to use
it whenever we're building new artifacts.

I've lifted and shifted the code we already had, and use a Dockerfile
type action. It only has a single gem as a dependency, so hopefully will
be relatively quick to build.