Check that correct architectures actually exist
I came across some interesting behaviour where we built a container from
a 3rd party container, where we only used the `FROM` line.

An upstream arm container wasn't actually available, and so an x86_64
container was used and pushed, but tagged with `arm64`.

This has surprising results in a CI pipeline further down the line,
since a manifest for that architecture did not exist.

Instead, let's protect against that by catching it during the manifest
push stage.