Remove bash set opts
This was causing the script to fail if a SLACK_WEBHOOK value was set, so
probably there is something else in the script which is failing. For our
purposes though we just want to cleanly exit if there is no Slack
webhook set, so I am not debugging this now.